287124 - made changes as per the review comments
