+++ README.md
@@ -104,15 +104,13 @@ npm run start:dev
 
 ```
 ├── src                    // Keeps the application code
-│   ├── controllers        // The definition of all route handlers can be found here
-│   │   └── web            // For web view rendering, controllers are arranged by module names
 │   ├── assets             // Contains all the sass styles and scripts
 │   │   ├── sass           // Contains all the sass styles and functions
 │   │   └── scripts        // Contains all the javascript modules or functions
-│   ├── utils              // utils contains common utility and helper reusable methods that can be used by controllers
-│   ├── services           // The definition of api clients and api handlers can be found here
-│   │   └── config         // To define and initiate the api client and interceptors
-│   │   └── handlers       // Contains all the API handlers, each endpoint will have its asynchronous function
+│   ├── config             // To define and initiate the configurations
+│   │   └── geoNetwork..ts // Here is where you can find the initialization of the axios clients
+│   ├── controllers        // The definition of all route handlers can be found here
+│   │   └── web            // For web view rendering, controllers are arranged by module names
 │   ├── infrastructure     // Frameworks, drivers, and plugins, to name a few.
 │   │   ├── log_files      // Contains the logs generated by the pino
 │   │   ├── plugins        // Contains the custom plugins registered with the Hapi server
@@ -122,6 +120,10 @@ npm run start:dev
 │   ├── routes             // Here is where all app routes are defined
 │   │   └── web            // WEB UI routes.
 │   │   │   └── assets.ts  // This is where all static asset routing requests are defined
+│   ├── services           // The definition of api clients and api handlers can be found here
+│   │   └── config         // To define and initiate the api client and interceptors
+│   │   └── handlers       // Contains all the API handlers, each endpoint will have its asynchronous
+│   ├── utils              // utils contains common utility and helper reusable methods that can be used by controllers function
 │   ├── views              // All server-rendered nunjucks templates, partials and helpers
 ├── └── index.ts           // The primary application entry point
 ├── public                 // Contains all the static files and files compiles from GDS
